Moved the logic to appActions. Updated the logic for updating an app.
